requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
Pillow>=10.0.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
gtts>=2.3.0
pygame>=2.5.0
python-socketio>=5.8.0
nltk>=3.8.0
flask>=2.3.0
flask-socketio>=5.3.0
eventlet>=0.33.0
